Create finance calculator script with monthly savings and annual projection

+++ python_introduction/finance_calculator.py
@@ -1,10 +1,10 @@
 #Declaration of variables
-monthly_income = float(input("Enter your monthly income: "))
-total_monthly_expenses = float(input("Enter your total monthly expenses: "))
+monthly_income = int(input("Enter your monthly income: "))
+total_monthly_expenses = int(input("Enter your total monthly expenses: "))
 rate = 0.05
 
 #Performing calculations
-monthly_savings = int(monthly_income - total_monthly_expenses)
+monthly_savings = monthly_income - total_monthly_expenses
 projected_savings = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
 
 #Outputing results
